requirements.txt
alembic
blinker
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.4
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.16
tomli==2.0.1
typing-extensions==4.6.3
urllib3==2.0.3
waitress==2.1.2
Werkzeug==2.3.6
zipp==3.15.0
Flask-Login